Silent request refactoring. MSAL will now check local cache for an access token instead of going directly to broker first
